apply linter
