+++ src/components/ui/ThemeChangeButton.tsx
@@ -1,5 +1,5 @@
 // import BaseButton from "./BaseButton";
-import React, { useContext } from 'react';
+import { useContext } from 'react';
 import { ThemeContext } from '../../context/ThemeContext';
 
 const ThemeChangeButton = () => {
